Refactoer: split pathpol tests into multiple files
Moves ACL tests to `act_test.go` and sequence tests to `sequence_test.go`
