+++ go/lib/pathpol/BUILD.bazel
@@ -27,6 +27,7 @@ go_test(
         "acl_test.go",
         "hop_pred_test.go",
         "policy_test.go",
+        "sequence_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
@@ -37,5 +38,6 @@ go_test(
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
+        "@org_golang_x_xerrors//:go_default_library",
     ],
 )
