+++ go/lib/pathpol/policy_test.go
@@ -58,406 +58,6 @@ func TestBasicPolicy(t *testing.T) {
 	}
 }
 
-func TestSequenceEval(t *testing.T) {
-	tests := map[string]struct {
-		Seq        *Sequence
-		Src        addr.IA
-		Dst        addr.IA
-		ExpPathNum int
-	}{
-		"Empty path": {
-			Seq:        newSequence(t, "0-0#0"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:212"),
-			ExpPathNum: 0,
-		},
-		"Length not matching": {
-			Seq:        newSequence(t, "0-0#0"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 0,
-		},
-		"Two Wildcard matching": {
-			Seq:        newSequence(t, "0-0#0 0-0#0"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"Longer Wildcard matching": {
-			Seq:        newSequence(t, "0-0#0 0-0#0 0-0#0 0-0#0"),
-			Src:        xtest.MustParseIA("1-ff00:0:122"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 2,
-		},
-		"Two Explicit matching": {
-			Seq:        newSequence(t, "1-ff00:0:133#1019 1-ff00:0:132#1910"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:132"),
-			ExpPathNum: 1,
-		},
-		"AS double IF matching": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#1910,1916 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 1,
-		},
-		"AS IF matching, first wildcard": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#0,1916 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 1,
-		},
-		"Longer Explicit matching": {
-			Seq: newSequence(t, "1-ff00:0:122#1815 1-ff00:0:121#1518,1530 "+
-				"1-ff00:0:120#3015,3122 2-ff00:0:220#2231,2224 2-ff00:0:221#2422"),
-			Src:        xtest.MustParseIA("1-ff00:0:122"),
-			Dst:        xtest.MustParseIA("2-ff00:0:221"),
-			ExpPathNum: 1,
-		},
-		"Longer Explicit matching, single wildcard": {
-			Seq: newSequence(t, "1-ff00:0:133#1018 1-ff00:0:122#1810,1815 "+
-				"1-ff00:0:121#0,1530 1-ff00:0:120#3015,2911 1-ff00:0:110#1129"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:110"),
-			ExpPathNum: 1,
-		},
-		"Longer Explicit matching, reverse single wildcard": {
-			Seq: newSequence(t, "1-ff00:0:133#1018 1-ff00:0:122#1810,1815 "+
-				"1-ff00:0:121#1530,0 1-ff00:0:120#3015,2911 1-ff00:0:110#1129"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:110"),
-			ExpPathNum: 0,
-		},
-		"Longer Explicit matching, multiple wildcard": {
-			Seq: newSequence(t, "1-ff00:0:133#1018 1-ff00:0:122#0,1815 "+
-				"1-ff00:0:121#0,1530 1-ff00:0:120#3015,0 1-ff00:0:110#1129"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:110"),
-			ExpPathNum: 1,
-		},
-		"Longer Explicit matching, mixed wildcard types": {
-			Seq: newSequence(t, "1-ff00:0:133#0 1 "+
-				"0-0#0 1-ff00:0:120#0 1-ff00:0:110#1129"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:110"),
-			ExpPathNum: 1,
-		},
-		"Longer Explicit matching, mixed wildcard types, two paths": {
-			Seq: newSequence(t, "1-ff00:0:133#0 1-0#0 "+
-				"0-0#0 1-0#0 1-ff00:0:110#0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:110"),
-			ExpPathNum: 2,
-		},
-		"Nil sequence does not filter": {
-			Seq:        nil,
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"Asterisk matches multiple hops": {
-			Seq:        newSequence(t, "0*"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"Asterisk matches zero hops": {
-			Seq:        newSequence(t, "0 0 0*"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"Plus matches multiple hops": {
-			Seq:        newSequence(t, "0+"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"Plus doesn't match zero hops": {
-			Seq:        newSequence(t, "0 0 0+"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 0,
-		},
-		"Question mark matches zero hops": {
-			Seq:        newSequence(t, "0 0 0?"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"Question mark matches one hop": {
-			Seq:        newSequence(t, "0 0?"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"Question mark doesn't match two hops": {
-			Seq:        newSequence(t, "0?"),
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 0,
-		},
-		"Successful match on hop count": {
-			Seq:        newSequence(t, "0 0 0"),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 3,
-		},
-		"Failed match on hop count": {
-			Seq:        newSequence(t, "0 0"),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 0,
-		},
-		"Select one of the intermediate ASes": {
-			Seq:        newSequence(t, "0 2-ff00:0:221 0"),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 1,
-		},
-		"Select two alternative intermediate ASes": {
-			Seq:        newSequence(t, "0 (2-ff00:0:221 | 2-ff00:0:210) 0"),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 3,
-		},
-		"Alternative intermediate ASes, but one doesn't exist": {
-			Seq:        newSequence(t, "0 (2-ff00:0:221 |64-12345) 0"),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 1,
-		},
-		"Or has higher priority than concatenation": {
-			Seq:        newSequence(t, "0 2-ff00:0:221|64-12345 0"),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 1,
-		},
-		"Question mark has higher priority than concatenation": {
-			Seq:        newSequence(t, "0 0 0 ?  "),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 3,
-		},
-		"Parentheses change priority": {
-			Seq:        newSequence(t, "(0 0)?"),
-			Src:        xtest.MustParseIA("2-ff00:0:211"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 0,
-		},
-		"Single interface matches inbound interface": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#1910 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 1,
-		},
-		"Single interface matches outbound interface": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#1916 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 1,
-		},
-		"Single non-matching interface": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#1917 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 0,
-		},
-		"Left interface matches inbound": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#1910,0 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 1,
-		},
-		"Left interface doesn't match outbound": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#1916,0 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 0,
-		},
-		"Right interface matches outbound": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#0,1916 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 1,
-		},
-		"Right interface doesn't match inbound": {
-			Seq:        newSequence(t, "0 1-ff00:0:132#0,1910 0"),
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("1-ff00:0:131"),
-			ExpPathNum: 0,
-		},
-	}
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	pp := NewPathProvider(ctrl)
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			paths := pp.GetPaths(test.Src, test.Dst)
-			outPaths := test.Seq.Eval(paths)
-			assert.Equal(t, test.ExpPathNum, len(outPaths))
-		})
-	}
-}
-
-var allowEntry = &ACLEntry{ACLAction(true), NewHopPredicate()}
-var denyEntry = &ACLEntry{ACLAction(false), NewHopPredicate()}
-
-func TestACLEval(t *testing.T) {
-	tests := map[string]struct {
-		ACL        *ACL
-		Src        addr.IA
-		Dst        addr.IA
-		ExpPathNum int
-	}{
-		"allow everything": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Allow, Rule: mustHopPredicate(t, "0-0#0")},
-					denyEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"allow 2-0#0, deny rest": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Allow, Rule: mustHopPredicate(t, "2-0#0")},
-					denyEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"allow 2-ff00:0:212#0 and 2-ff00:0:211, deny rest": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Allow, Rule: mustHopPredicate(t, "2-ff00:0:212#0")},
-					{Action: Allow, Rule: mustHopPredicate(t, "2-ff00:0:211#0")},
-					denyEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 2,
-		},
-		"allow 2-ff00:0:212#0, deny rest": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Allow, Rule: mustHopPredicate(t, "2-ff00:0:212#0")},
-					denyEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("2-ff00:0:212"),
-			Dst:        xtest.MustParseIA("2-ff00:0:211"),
-			ExpPathNum: 0,
-		},
-		"deny 1-ff00:0:110#0, 1-ff00:0:120#0, allow rest": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Deny, Rule: mustHopPredicate(t, "1-ff00:0:110#0")},
-					{Action: Deny, Rule: mustHopPredicate(t, "1-ff00:0:120#0")},
-					allowEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("2-ff00:0:222"),
-			ExpPathNum: 2,
-		},
-		"deny 1-ff00:0:110#0, 1-ff00:0:120#0 and 1-ff00:0:111#2823, allow rest": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Deny, Rule: mustHopPredicate(t, "1-ff00:0:110#0")},
-					{Action: Deny, Rule: mustHopPredicate(t, "1-ff00:0:120#0")},
-					{Action: Deny, Rule: mustHopPredicate(t, "1-ff00:0:111#2823")},
-					allowEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("1-ff00:0:133"),
-			Dst:        xtest.MustParseIA("2-ff00:0:222"),
-			ExpPathNum: 1,
-		},
-		"deny ISD1, allow certain ASes": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Allow, Rule: mustHopPredicate(t, "1-ff00:0:120#0")},
-					{Action: Allow, Rule: mustHopPredicate(t, "1-ff00:0:130#0")},
-					{Action: Deny, Rule: mustHopPredicate(t, "1-0#0")},
-					allowEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("1-ff00:0:130"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 2,
-		},
-		"deny ISD1, allow certain ASes - wrong oder": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Deny, Rule: mustHopPredicate(t, "1-0#0")},
-					{Action: Allow, Rule: mustHopPredicate(t, "1-ff00:0:130#0")},
-					{Action: Allow, Rule: mustHopPredicate(t, "1-ff00:0:120#0")},
-					allowEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("1-ff00:0:130"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 0,
-		},
-		"nil rule should match all the paths": {
-			ACL: &ACL{
-				Entries: []*ACLEntry{
-					{Action: Deny, Rule: nil},
-					allowEntry,
-				},
-			},
-			Src:        xtest.MustParseIA("1-ff00:0:130"),
-			Dst:        xtest.MustParseIA("2-ff00:0:220"),
-			ExpPathNum: 0,
-		},
-	}
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	pp := NewPathProvider(ctrl)
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			paths := pp.GetPaths(test.Src, test.Dst)
-			outPaths := test.ACL.Eval(paths)
-			assert.Equal(t, test.ExpPathNum, len(outPaths))
-		})
-	}
-}
-
-func TestACLPanic(t *testing.T) {
-	acl := &ACL{
-		Entries: []*ACLEntry{
-			{Action: Allow, Rule: mustHopPredicate(t, "1-0#0")},
-		},
-	}
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	pp := NewPathProvider(ctrl)
-	paths := pp.GetPaths(xtest.MustParseIA("2-ff00:0:212"), xtest.MustParseIA("2-ff00:0:211"))
-	assert.Panics(t, func() { acl.Eval(paths) })
-}
-
-func TestACLConstructor(t *testing.T) {
-	_, err := NewACL(&ACLEntry{Action: Allow, Rule: mustHopPredicate(t, "1-0#0")})
-	if assert.Error(t, err) {
-		assert.Equal(t, ErrNoDefault, err)
-	}
-	acl, err := NewACL(
-		&ACLEntry{Action: Allow, Rule: mustHopPredicate(t, "1-0#0")},
-		&ACLEntry{Action: Deny, Rule: mustHopPredicate(t, "0-0#0")},
-	)
-	if assert.NoError(t, err) {
-		assert.NotNil(t, acl)
-	}
-}
-
 func TestOptionsEval(t *testing.T) {
 	tests := map[string]struct {
 		Policy     *Policy
@@ -981,22 +581,6 @@ func TestFilterOpt(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			outPaths := test.Policy.FilterOpt(paths, FilterOptions{IgnoreSequence: true})
 			assert.Equal(t, test.ExpPathNum, len(outPaths))
-		})
-	}
-}
-
-func TestSequenceConstructor(t *testing.T) {
-	tests := map[string]assert.ErrorAssertionFunc{
-		"0-0-0#0": assert.Error,
-		"0#0#0":   assert.Error,
-		"0":       assert.NoError,
-		"1#0":     assert.Error,
-		"1-0":     assert.NoError,
-	}
-	for seq, assertion := range tests {
-		t.Run(seq, func(t *testing.T) {
-			_, err := NewSequence(seq)
-			assertion(t, err, seq)
 		})
 	}
 }
